instant-rag-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
asyncpg==0.28.0
psycopg2-binary==2.9.9
pgvector==0.2.3
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
email-validator==2.1.0
aiofiles==23.2.1
sentence-transformers==2.2.2
huggingface-hub==0.19.4
numpy==1.26.1
aiohttp==3.8.6
python-dotenv==1.0.0
PyYAML==6.0.1
PyMuPDF==1.23.7
pytesseract==0.3.10
Pillow==10.1.0
tiktoken==0.5.1
cryptography==41.0.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
langchain==0.2.0
langchain-core==0.2.0
langchain-community==0.2.0
langchain-text-splitters==0.2.0
jsonschema==4.19.0
httpx==0.25.0
beautifulsoup4==4.12.2
python-docx==0.8.11
